Revert "Extract thunder_gqls test base class functionality into a trait"
This reverts commit 04ccb6997107819f22c1c2e6430beea93f21761b.
