+++ modules/thunder_gqls/tests/src/Functional/JsonLdSchemaTest.php
@@ -20,6 +20,7 @@ class JsonLdSchemaTest extends ThunderGqlsTestBase {
     $this->config('metatag.metatag_defaults.node__article')->set('tags', $tags)
       ->save();
 
+    $this->drupalLogin($this->graphqlUser);
     $this->runAndTestQuery('jsonld');
   }
